// src/app/auth/set-new-password/set-new-password-enums.enum.ts
export enum PasswordValidationMessages {
    PleaseEnterNewPassword = "PleaseEnterNewPassword",
    PleaseConfirmNewPassword = "PleaseConfirmNewPassword",
    MinimumCharactersLimitOfPassword = "MinimumCharacterslimitofPassword",
    MaximumCharactersLimitOfPassword = "MaximumCharacterslimitofPassword"
}

export enum ToastMessages {
    YourPasswordHasBeenChangedSuccessfully = "Your password has been changed successfully."
}